requirements.txt
python-telegram-bot==21.6
google-genai==1.3.0
sqlalchemy==2.0.36
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.11.11
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
